requirements.txt
pandas
numpy
duckdb
sqlalchemy
langchain==0.1.16
langchain-community==0.0.32
langchain-google-genai==1.0.3
sentence-transformers
streamlit
scikit-learn
matplotlib
seaborn
tqdm
python-dotenv
faiss-cpu
duckdb-engine
tabulate
wikipedia
tavily
googletrans